@@ -361,11 +361,11 @@ export class LearningObjectService {
   }
 
   /**
-   * Fetchs the parents of a learning object
-   * @param id of learing object
+   * Fetches the parents of a learning object
+   * @param id of learning object
    */
-  fetchParents(id: string) {
-    const route = PUBLIC_LEARNING_OBJECT_ROUTES.GET_LEARNING_OBJECT_PARENTS(id);
+  fetchParents(username: string, id: string) {
+    const route = PUBLIC_LEARNING_OBJECT_ROUTES.GET_LEARNING_OBJECT_PARENTS(username, id);
     return this.http.get<LearningObject[]>(route, { withCredentials: true }).toPromise().then(parents => {
       return parents;
     });
