@@ -4,13 +4,22 @@ import * as querystring from 'querystring';
 export type MaterialsFilter = 'released' | 'unreleased';
 
 export const ADMIN_ROUTES = {
-  MUTATE_COLLECTION_MEMBERSHIP(abvCollectionName: string, userId: string) {
+  MUTATE_COLLECTION_MEMBERSHIP(abvCollectionName: string, userId: string): string {
     return `${environment.apiURL}/collections/${encodeURIComponent(
       abvCollectionName
     )}/members/${encodeURIComponent(userId)}`;
   },
-  GET_USER_ROLES(id: string) {
+  GET_USER_ROLES(id: string): string {
     return `${environment.apiURL}/users/${encodeURIComponent(id)}/roles`;
+  },
+  GET_MAPPERS(): string {
+    return `${environment.apiURL}/guidelines/members`;
+  },
+  ADD_MAPPER(userId: string): string {
+    return `${environment.apiURL}/guidelines/members/${encodeURIComponent(userId)}`;
+  },
+  REMOVE_MAPPER(userId: string): string {
+    return `${environment.apiURL}/guidelines/members/${encodeURIComponent(userId)}`;
   }
 };
 
@@ -379,8 +388,8 @@ export const PUBLIC_LEARNING_OBJECT_ROUTES = {
       username
     )}/learning-objects`;
   },
-  GET_LEARNING_OBJECT_PARENTS(id: string) {
-    return `${environment.apiURL}/learning-objects/${id}/parents`;
+  GET_LEARNING_OBJECT_PARENTS(username: string, id: string) {
+    return `${environment.apiURL}/users/${username}/learning-objects/${id}/parents`;
   },
   DOWNLOAD_FILE(params: {
     username: string;
